Separate player and enemy attack, rebalance to have smaller cooldowns and better damage counts / ranges

+++ Assets/Scripts/Game/Data/AbilityDatabase.cs
@@ -24,8 +24,8 @@ namespace Game.Data
             public CharmCost CharmCost => m_charmCost;
 
             [SerializeField]
-            private int m_cooldownSeconds = 6;
-            public int CooldownSeconds => m_cooldownSeconds;
+            private float m_cooldownSeconds = 6;
+            public float CooldownSeconds => m_cooldownSeconds;
 
             [SerializeField]
             private Sprite m_uiImage = null;
